requirements.txt
requests
beautifulsoup4
unidecode
pandas
python-dotenv
tqdm
